Checkpoints now use a .done file; job_stream can specify checkpoints.
C++ tests may not be working, this computer has bad GCC version.

Version bumped to 0.1.18 for Python.

+++ python/job_stream/test/common.py
@@ -86,7 +86,13 @@ class JobStreamTest(unittest.TestCase):
             return self.execute(f.name, np=np)
 
 
-    def safeRemove(self, path):
+    def safeRemove(self, pathOrPaths):
+        path = pathOrPaths
+        if isinstance(path, (tuple, list)):
+            for p in path:
+                self.safeRemove(p)
+            return
+
         try:
             os.remove(path)
         except OSError as e:
